HPCC-18009 Move WorkunitErrorReceiver out of hql
Allowing hql.dll to be used more widely with lower footprint.
