+++ common/workunit/workunit.cpp
@@ -10839,3 +10839,35 @@ void addWorkunitException(IWorkUnit * wu, IError * error, bool removeTimeStamp)
     if (error->queryScope())
         exception->setScope(error->queryScope());
 }
+
+
+IError * WorkUnitErrorReceiver::mapError(IError * error)
+{
+    return LINK(error);
+}
+
+void WorkUnitErrorReceiver::report(IError* eclError)
+{
+    addWorkunitException(wu, eclError, removeTimeStamp);
+}
+
+size32_t WorkUnitErrorReceiver::errCount()
+{
+    unsigned count = 0;
+    Owned<IConstWUExceptionIterator> exceptions = &wu->getExceptions();
+    ForEach(*exceptions)
+        if (exceptions->query().getSeverity() == SeverityError)
+            count++;
+    return count;
+}
+
+size32_t WorkUnitErrorReceiver::warnCount()
+{
+    unsigned count = 0;
+    Owned<IConstWUExceptionIterator> exceptions = &wu->getExceptions();
+    ForEach(*exceptions)
+        if (exceptions->query().getSeverity() == SeverityWarning)
+            count++;
+    return count;
+}
+
