+++ common/workunit/workunit.hpp
@@ -1521,6 +1521,25 @@ extern WORKUNIT_API WUAction getWorkunitAction(const char * actionStr);
 
 extern WORKUNIT_API void addTimeStamp(IWorkUnit * wu, StatisticScopeType scopeType, const char * scope, StatisticKind kind);
 extern WORKUNIT_API IPropertyTree * getWUGraphProgress(const char * wuid, bool readonly);
+
+class WORKUNIT_API WorkUnitErrorReceiver : implements IErrorReceiver, public CInterface
+{
+public:
+    WorkUnitErrorReceiver(IWorkUnit * _wu, const char * _component, bool _removeTimeStamp) { wu.set(_wu); component.set(_component); removeTimeStamp = _removeTimeStamp; }
+    IMPLEMENT_IINTERFACE;
+
+    virtual IError * mapError(IError * error);
+    virtual void exportMappings(IWorkUnit * wu) const { }
+    virtual void report(IError*);
+    virtual size32_t errCount();
+    virtual size32_t warnCount();
+
+private:
+    Owned<IWorkUnit> wu;
+    StringAttr component;
+    bool removeTimeStamp;
+};
+
 extern WORKUNIT_API void addWorkunitException(IWorkUnit * wu, IError * error, bool removeTimeStamp);
 
 inline bool isGlobalScope(const char * scope) { return scope && (streq(scope, GLOBAL_SCOPE) || streq(scope, LEGACY_GLOBAL_SCOPE)); }
